Reorganized integration tests
